Mon 3:50PM

+++ database/schema.sql
@@ -19,7 +19,7 @@ CREATE TABLE app_user (
 
 CREATE TABLE pothole (
   id SERIAL PRIMARY KEY,
-  address_id int(32) NOT NULL,
+  address_id long NOT NULL,
   size varchar(32) NOT NULL,
   description varchar(32),
   created_on date(255) NOT NULL,
@@ -29,7 +29,7 @@ CREATE TABLE pothole (
 
 CREATE TABLE address (
   address_id SERIAL PRIMARY KEY,
-  zip_code int(32) NOT NULL,
+  zip_code int NOT NULL,
   city varchar(32) NOT NULL,
   address_line_1 varchar (255) NOT NULL,
   address_line_2 varchar (255) NOT NULL
